+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+xxx
+===================
+- Added `Apify.utils.htmlToText()` function to convert HTML to text and removed unncessary `html-to-text` dependency.
+  The new function is now used in `Apify.utils.social.parseHandlesFromHtml()`.
+
 0.10.0 / 2018-12-19
 ===================
 - `autoscaledPool.isFinishedFunction()` and `autoscaledPool.isTaskReadyFunction()` exceptions